+++ psoc5/firmware/PSOC_trigger_firmware_Trigger_Polarity_Low/Trigger.cydsn/Generated_Source/PSoC5/project.h
@@ -39,7 +39,7 @@
 #include "freqcounter.h"
 #include "Timer.h"
 #include "muxSel.h"
-#include "DMA_dma.h"
+#include "isr_timeStamp.h"
 #include "UART_IntClock.h"
 #include "core_cm3_psoc5.h"
 #include "CyDmac.h"